Stop importing as `index.js` to make `esm.sh` happy
It seems that `esm.sh` strips `index.js` from the import path while
bundling. This causes wrong `import "dist.js"` to be generated.

Close https://github.com/ruby/ruby.wasm/issues/335
